Split units for option adapter
