+++ tests/unit/plugins/plugin_utils/base/test_hashi_vault_plugin.py
